Break out generic structure store
