+++ assets/blocks/course-outline/store.js
@@ -1,30 +1,23 @@
 /**
  * External dependencies
  */
-import { isEqual } from 'lodash';
 
 /**
  * WordPress dependencies
  */
-import { apiFetch, controls as dataControls } from '@wordpress/data-controls';
-import { dispatch, registerStore, select, subscribe } from '@wordpress/data';
+import { dispatch, select } from '@wordpress/data';
 import { __, sprintf } from '@wordpress/i18n';
 
 /**
  * Internal dependencies
  */
-import { createReducerFromActionMap } from '../../shared/data/store-helpers';
+import { registerStructureStore } from '../../shared/structure/structure-store';
+
 import {
 	syncStructureToBlocks,
 	extractStructure,
 	getFirstBlockByName,
 } from './data';
-
-const DEFAULT_STATE = {
-	serverStructure: null,
-	isSavingStructure: false,
-	hasStructureUpdate: false,
-};
 
 const getEditorOutlineBlock = () =>
 	getFirstBlockByName(
@@ -42,81 +35,15 @@ const getEditorOutlineStructure = () => {
 	return extractStructure( outlineBlock.innerBlocks );
 };
 
-const actions = {
-	/**
-	 * Fetch course structure data from REST API.
-	 */
-	*fetchCourseStructure() {
+export const COURSE_STORE = 'sensei/course-structure';
+
+registerStructureStore( {
+	storeName: COURSE_STORE,
+	*getEndpoint() {
 		const courseId = yield select( 'core/editor' ).getCurrentPostId();
-		const result = yield apiFetch( {
-			path: `/sensei-internal/v1/course-structure/${ courseId }?context=edit`,
-		} );
-		yield actions.setStructure( result );
+		return `course-structure/${ courseId }`;
 	},
-
-	/**
-	 * Persist editor's course structure to the REST API.
-	 *
-	 * @param {Array} editorStructure
-	 */
-	*saveStructure( editorStructure ) {
-		yield { type: 'SAVING', isSavingStructure: true };
-		const courseId = yield select( 'core/editor' ).getCurrentPostId();
-
-		try {
-			const result = yield apiFetch( {
-				path: `/sensei-internal/v1/course-structure/${ courseId }`,
-				method: 'POST',
-				data: { structure: editorStructure },
-			} );
-			yield actions.setStructure( result, editorStructure );
-		} catch ( error ) {
-			const errorMessage = sprintf(
-				/* translators: Error message. */
-				__(
-					'Course modules and lessons could not be updated. %s',
-					'sensei-lms'
-				),
-				error.message
-			);
-			yield dispatch( 'core/notices' ).createErrorNotice( errorMessage, {
-				id: 'course-outline-save-error',
-			} );
-		}
-
-		yield { type: 'SAVING', isSavingStructure: false };
-	},
-
-	/**
-	 * Set fetched structure.
-	 *
-	 * @param {Array} serverStructure
-	 * @param {Array} editorStructure
-	 */
-	*setStructure( serverStructure, editorStructure = null ) {
-		yield actions.setServerStructure( serverStructure, editorStructure );
-		yield actions.updateOutlineBlock( serverStructure );
-	},
-
-	/**
-	 * Keep last fetched server state for comparison.
-	 *
-	 * @param {Array} serverStructure
-	 * @param {Array} editorStructure
-	 */
-	setServerStructure: ( serverStructure, editorStructure = null ) => ( {
-		type: 'SET_SERVER_STRUCTURE',
-		serverStructure,
-		hasStructureUpdate:
-			editorStructure && ! isEqual( serverStructure, editorStructure ),
-	} ),
-
-	/**
-	 * Update outline block.
-	 *
-	 * @param {Array} structure
-	 */
-	*updateOutlineBlock( structure ) {
+	*updateBlock( structure ) {
 		const { clientId = null } = getEditorOutlineBlock();
 
 		if ( ! clientId || ! structure || 0 === structure.length ) {
@@ -132,109 +59,22 @@ const actions = {
 			false
 		);
 	},
-
-	/**
-	 * Clear structure update.
-	 */
-	clearStructureUpdate: () => ( { type: 'CLEAR_STRUCTURE_UPDATE' } ),
-};
-
-/**
- * Course structure reducers.
- */
-const reducers = {
-	SET_SERVER_STRUCTURE: (
-		{ serverStructure, hasStructureUpdate },
-		state
-	) => {
-		return {
-			...state,
-			serverStructure,
-			hasStructureUpdate,
-		};
+	readBlock: getEditorOutlineStructure,
+	*saveError( error ) {
+		const errorMessage = sprintf(
+			/* translators: Error message. */
+			__(
+				'Course modules and lessons could not be updated. %s',
+				'sensei-lms'
+			),
+			error.message
+		);
+		yield dispatch( 'core/notices' ).createErrorNotice( errorMessage, {
+			id: 'course-outline-save-error',
+		} );
 	},
-	SAVING: ( { isSavingStructure }, state ) => ( {
-		...state,
-		isSavingStructure,
-	} ),
-	CLEAR_STRUCTURE_UPDATE: ( action, state ) => ( {
-		...state,
-		hasStructureUpdate: false,
-	} ),
-	DEFAULT: ( action, state ) => state,
-};
-
-/**
- * Course structure selectors.
- */
-const selectors = {
-	shouldResavePost: ( { hasStructureUpdate } ) => hasStructureUpdate,
-	getIsSavingStructure: ( { isSavingStructure } ) => isSavingStructure,
-	getServerStructure: ( { serverStructure } ) => serverStructure,
-};
-
-export const COURSE_STORE = 'sensei/course-structure';
-
-/**
- * Register course structure store and subscribe to block editor save.
- */
-const registerCourseStructureStore = () => {
-	// Set to true when savings starts, and false when it ends.
-	let postSaving = false;
-
-	const startSave = () => {
-		const serverStructure = select( COURSE_STORE ).getServerStructure();
-		const editorStructure = getEditorOutlineStructure();
-
-		if (
-			! editorStructure ||
-			isEqual( serverStructure, editorStructure )
-		) {
-			return;
-		}
-
+	clearError() {
 		// Clear error notices.
 		dispatch( 'core/notices' ).removeNotice( 'course-outline-save-error' );
-		dispatch( COURSE_STORE ).saveStructure( editorStructure );
-	};
-
-	const finishSave = () => {
-		const shouldResavePost = select( COURSE_STORE ).shouldResavePost();
-
-		if ( ! shouldResavePost ) {
-			return;
-		}
-
-		dispatch( 'core/editor' ).savePost();
-		dispatch( COURSE_STORE ).clearStructureUpdate();
-	};
-
-	subscribe( function saveStructureOnPostSave() {
-		const editor = select( 'core/editor' );
-
-		if ( ! editor ) return;
-
-		const isSavingPost =
-			editor.isSavingPost() && ! editor.isAutosavingPost();
-		const isSavingStructure = select( COURSE_STORE ).getIsSavingStructure();
-
-		if ( ! postSaving && isSavingPost ) {
-			// First update where post is saving.
-			postSaving = true;
-			startSave();
-		} else if ( postSaving && ! isSavingPost && ! isSavingStructure ) {
-			// First update where post is no longer saving and editor is sync.
-			postSaving = false;
-			finishSave();
-		}
-	} );
-
-	registerStore( COURSE_STORE, {
-		reducer: createReducerFromActionMap( reducers, DEFAULT_STATE ),
-		actions,
-		selectors,
-		controls: { ...dataControls },
-	} );
-};
-
-registerCourseStructureStore();
+	},
+} );
